requirements.txt
dash==2.14.1
dash-bootstrap-components==1.5.0
dash-leaflet==0.1.23
plotly==5.15.0
pandas==2.0.3
meteomatics-api==1.0.5
python-dotenv==1.0.0